se actualizó la clase clienteService convirtiendola en una interfaz que extiende de CrudGeneric, esto para poder agregarle métodos específicos de ser necesario, la clase clienteServiceImp ahora implementa la interfaz ClienteService, de esta manera el endpoint de /cliente cumple con todas las peticiones del crud

+++ src/main/java/ar/com/peluqueriacanina/generic/CrudGeneric.java
@@ -4,11 +4,11 @@ import java.util.List;
 
 public interface CrudGeneric <T>{
 	
-	   public T findById(Long id);
+	   public T findById(int id);
 	   public T save(T model);
-	   public List<T> findAll(Long id);
+	   public List<T> findAll(int id);
 	   public List<T> findAll();
-	   public Boolean delete(Long id);
+	   public Boolean delete(int id);
 	   public T update(T model);
